almost ready

+++ CrooksAndCastles/CrooksAndCastles/Interfaces/ISkills.cs
@@ -4,8 +4,8 @@ namespace DrunkenSoftUniWarrior.Interfaces
 {
     public interface ISkills
     {
-        int Health { get; set; }
-        int Damage { get; set; }
+        double Health { get; set; }
+        double Damage { get; set; }
         bool IsAlive { get; set; }
     }
 }
